Reorganized screens
Reorganized screens in preparation for Case Management navigation etc.

Jira ticket: CAMS-219
