+++ user-interface/src/case-assignment/AssignAttorneyModal.scss
