+++ user-interface/src/case-assignment/CaseAssignment.scss
