+++ user-interface/src/case-detail/CaseDetailScreen.scss
