+++ user-interface/src/lib/components/HeaderNavBar.scss
@@ -1,4 +1,4 @@
-@use '../styles/abstracts/colors';
+@use '@/styles/abstracts/colors';
 
 .usa-header > * {
   background-color: colors.$navy;
