+++ user-interface/src/lib/components/HeaderNavBar.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
-import App from '../App';
+import App from '../../App';
 
 describe('App', () => {
   it('loads a page', async () => {
