+++ user-interface/src/lib/components/LoadingIndicator.tsx
@@ -1,0 +1,3 @@
+export default function LoadingIndicator() {
+  return <p data-testid="loading-indicator">Loading...</p>;
+}
