+++ user-interface/src/lib/hooks/UseComponent.ts
