+++ user-interface/src/lib/hooks/UseOutsideClick.ts
