+++ user-interface/src/lib/models/attorneys-api.ts
@@ -1,5 +1,5 @@
 import Api from './api';
-import { AttorneyListResponseData } from '@/components/AssignAttorneyModal';
+import { AttorneyListResponseData } from '@/case-assignment/AssignAttorneyModal';
 import AttorneysMockApi from './attorneys-mock.api.cases';
 
 export default class AttorneysApi {
