+++ user-interface/src/lib/type-declarations/basic.d.ts
