+++ user-interface/src/lib/utils/formatCaseNumber.test.ts
