+++ user-interface/src/lib/utils/formatCaseNumber.ts
