+++ user-interface/src/lib/utils/http.adapter.ts
