+++ user-interface/src/lib/utils/mock.update-cases.ts
@@ -1,4 +1,4 @@
-import { CallBackProps } from '../components/AssignAttorneyModal';
+import { CallBackProps } from '../../case-assignment/AssignAttorneyModal';
 
 export default class MockUpdateCases {
   public static mockCallback(props: CallBackProps) {
